1.对齐双端的基本概念 Delegate都基于接口实现 2.梳理Boost的API以及必要注释，非必要的API放起来讨论 3.对齐OC的语法风格，干掉Java风格的OC代码
