+++ example/android/app/src/main/java/com/idlefish/flutterboost/example/MyApplication.java
@@ -18,30 +18,11 @@ public class MyApplication extends FlutterApplication {
     public void onCreate() {
         super.onCreate();
 
-        FlutterBoost.instance().setup(this, new FlutterBoostDelegate() {
-
-            @Override
-            public void pushNativeRoute(String pageName, HashMap<String, String> arguments) {
-                Intent intent = new Intent(FlutterBoost.instance().currentActivity(), NativePageActivity.class);
-                FlutterBoost.instance().currentActivity().startActivity(intent);
-            }
-
-            @Override
-            public void pushFlutterRoute(String pageName, String uniqueId, HashMap<String, String> arguments) {
-                Intent intent = new FlutterBoostActivity.CachedEngineIntentBuilder(FlutterBoostActivity.class, FlutterBoost.ENGINE_ID)
-                        .backgroundMode(FlutterActivityLaunchConfigs.BackgroundMode.opaque)
-                        .destroyEngineWithActivity(false)
-                        .uniqueId(uniqueId)
-                        .url(pageName)
-                        .urlParams(arguments)
-                        .build(FlutterBoost.instance().currentActivity());
-                FlutterBoost.instance().currentActivity().startActivity(intent);
-            }
-
-        },engine->{
+        FlutterBoost.instance().setup(this, new MyFlutterBoostDelegate(),engine->{
             engine.getPlugins();
         } );
 
 
     }
 }
+
